requirements.txt
Flask
Flask-SQLAlchemy
PyJWT
Werkzeug
requests
python-dotenv
